Remove `AuthorityClient`
* Remove `AuthorityClient` as it is no longer needed
* Add `hasUserRole` caller to `ColonyClient`
* Rename `OWNER` role to `FOUNDER`
