+++ packages/colony-js-client/src/constants.js
@@ -12,11 +12,11 @@ export const ROLES = {
 
 export const DEFAULT_DOMAIN_ID = 1;
 
-export const OWNER_ROLE = 'OWNER';
+export const FOUNDER_ROLE = 'FOUNDER';
 export const ADMIN_ROLE = 'ADMIN';
 
 export const AUTHORITY_ROLES = {
-  [OWNER_ROLE]: 0,
+  [FOUNDER_ROLE]: 0,
   [ADMIN_ROLE]: 1,
 };
